Rework ViewState (now Observable) to handle nested Observable objects
A property of an Observable is observed if and only if it has the Observed property
wrapper even if it is Observable (for clarity reasons). ViewState and AppState are
now just deprecated aliases to Observable. The Observed property wrapper itself
now conforms to Observable instead of PublishedValue (they were basically the same
protocol)