Impl @Environment property wrapper, impl file open/save dialog support
The @Environment property wrapper work should make @State much easier.
In this commit I also cleaned up both Gtk wrappers quite significantly
with a new root GObject class and a focus on convenience initializers
instead of designated initializers (which are both more correct and
simpler to generate due to not requiring us to determine which inits are
overrides and which aren't).

I would've done @Environment and file dialog support separately but I
only started on @Environment because it was required for file dialog
support, so it all got a bit jumbled together.