Fix Gtk3Backend button-press-event signal handler crash
The Gtk 3 docs say that the signal handler should take GdkEventButton
as a value, but in reality it seems that we have to take it as
a pointer. To reproduce the issue, double click any clickable view (e.g.
a TextField) in a Gtk3Backend app. The crash occurs because the
incorrect argument size causes the unmanaged SignalBox pointer to be
some random pointer. Why the crash didn't occur on non Rocky Linux
platforms still evades me.