Bring back the terminal backend to life.
It renames 'CursesBackend' to TermKit, as now TermKit has various
console drivers, and curses is just one of them.

The challenge is that TermKit deals with screen sizes like 80x24, not
things like 600x400, so both the default sizes and spacing in the
codebase will need to compensate for that.

The AppBackend introduces a defaultStackSpacingAmount which is the
spacing that VStack and HStack would use, it is currently kept as a
default to the existing value, but for the TermKit backend, this value
is set to zero.

Now HStack and VStack rather than storing a resolved value like
'spacing: Int', they store the request 'spacing: Int?', which is
resolved before it is actually needed by querying the backend for this
data.

There is also a new 'limitScreenBounds' that is a no-op for most, but
on the curses cases, ensures that a misuse of the defaultSize does not
trickle up and down the stack with unrealistic view sizes.