Fix layout caching bug: don't treat cached layout as 'current layout'
This led to situations where the last accessed cached layout would
replace the real current layout and then the layout system would commit
a cached layout instead of the current layout, leading to strange
results. I noticed this bug when testing StressTestExample. Open the
app and then click Generate without interacting with the window in any
way beforehand to reproduce the bug (resizing the window first seems to
stop the bug). I only tried reproducing the bug on macOS with
AppKitBackend.